Create a function that plays a single round
I created a function that plays a single round of Rock, Paper, Scissors
by converting the playerSelection and computerSelection variables
toUpperCase. It then compares the player and computer to see if they
match and end in a tie. If it is not a tie I then use if and else if
statements to compare the player and computer and determine a winner!